Move sentinel and webapp message-utils into shared-libs
Issue: #4420
