+++ shared-libs/message-utils/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "@shared-libs/message-utils",
+  "version": "1.0.0",
+  "main": "src/index.js",
+  "scripts": {
+    "test": "mocha ./test"
+  },
+  "devDependencies": {
+    "chai": "^4.1.2",
+    "mocha": "^5.1.0"
+  }
+}
