+++ shared-libs/message-utils/test/.jshintrc
@@ -1,0 +1,4 @@
+{
+  "esversion": 6,
+  "mocha": true
+}
